Add auth token to users
We should probably make a whole new table in the future, but that seemed
like overkill for this stage. My current plan is to just add a login
mutation that sets the auth token and a logout mutation that clears it.
It would probably be better to have user sessions or something or make
the token expire after a certain amount of time. But, I'm just trying to
get this thing out the door at this point. If you think this should be
changed in the future then by all means go ahead.