Remove user relation from graphql
First of all, the type should be user, not Project. Second, the way
typeorm eagerly loads the relationship isn't bi-directional. For that
reason it justifies the logic of "Start at a user and work down". So, I
don't feel like we really lose a lot hear not being able to traverse
back up the relationship.

Having worked on bigger projects I think having bi-directional relations
just makes things more complicated in the future. So I think this is
actually a good thing for now.