Add authChecker function
This is required for using the @Authorized directive in resolvers. This
is a very naive approach because a user can be an admin in one
organization, and a member in another. So to figure out if they are
authorized depends a lot on the endpiont. So this may not work long
term. I.e., if in the future this method is useless then remove it.